chore: Refactor the Problem component
The `Problem` component is all in one file, and has multiple components.
This giant pile of code is hard to read and maintain. This change moves
the subcomponents out to separate files so that they're easier to
maintain & traverse. Maybe we can even add tests and storybook story
definitions some day!
