+++ src/components/Problem/ProblemComments/index.ts
@@ -1,0 +1,1 @@
+export { ProblemComments } from "./ProblemComments";
