+++ src/components/Problem/ProblemTicks/index.ts
@@ -1,0 +1,1 @@
+export { ProblemTicks } from "./ProblemTicks";
