+++ src/components/Problem/ProblemsOnRock/index.ts
@@ -1,0 +1,1 @@
+export { ProblemsOnRock } from "./ProblemsOnRock";
