+++ src/components/Problem/index.ts
@@ -1,0 +1,1 @@
+export { Problem as default } from "./Problem";
